refactor(cockpit): extract _SettingsAccountReassignModal to sibling module (#1354)
Wedge 4 of the cockpit_ui.py god-module split. Moves
_SettingsAccountReassignModal into cockpit_settings_account_reassign.py
alongside the prior _SettingsConfirmModal extraction; cockpit_ui re-
exports the class so existing callers keep working.
